Fix issue 312. Now the label width field in gconf is in pixel rather than in chars.
Added max_width_chars to MenuItem and MenuLabel. The default max_width_chars
in Gnomenu is 30 chars. This member has no correspondance in XML, since it
is a pure visualization term.

Improved the accuracy of size allocation algorithm in MenuLabel.
Expanded the range of the Spinbutton in GConfDialog.
Fixed a possible core dump in MenuBar implementation.




git-svn-id: http://gnome2-globalmenu.googlecode.com/svn/trunk@2346 1ffc4f22-0e42-0410-a188-a74355021015